+++ headers/hardware/vga/vga_sequencer.h
@@ -73,7 +73,7 @@ typedef union PACKED
 } LOADEDPLANESCONTAINER; //All four loaded planes!
 #include "headers/endpacked.h"
 
-float VGA_VerticalRefreshRate(VGA_Type *VGA); //Scanline speed for one line in Hz!
+double VGA_VerticalRefreshRate(VGA_Type *VGA); //Scanline speed for one line in Hz!
 
 void VGA_Sequencer_calcScanlineData(VGA_Type *VGA);
